PLT-477 Profile tags for Personalised Content - Code reusability and Fix for multiple Answer in Submit
